refactor(server): consolidate matching engines — finalize enhanced, remove legacy files
